refactor(api): merge experiment network controller into network controller
